// alix-pwa/src/types/events.ts
export interface Events {
    id: number;
    published_at: string;
    start_time: string;
    end_time: string;
    communeId: number;
  
    title_1: string;
    title_2?: string;
    title_3?: string;
  
    text_1: string;
    text_2?: string;
    text_3?: string;
  
    img_1?: string;
    img_2?: string;
    img_3?: string;
  }
  